mqtt/requirements.txt
python-dotenv
paho-mqtt==2.0.0
requests
opencv-python
numpy
onnxruntime-gpu==1.15.1